// src/mocks/types.ts
export type CardStatsHorizontalType = {
    icon: string
    title: string
    stats: string
    color?: ThemeColor
    trendNumber: string
    trend?: 'positive' | 'negative'
}
export type CardStatsVerticalType = {
    icon: string
    title: string
    stats: string
    subtitle: string
    color?: ThemeColor
    trendNumber: string
    trend?: 'positive' | 'negative'
    optionsMenuProps?: OptionsMenuType
}
export type CardStatsCharacterType = {
    src: string
    title: string
    stats: string
    chipText: string
    trendNumber: string
    chipColor?: ThemeColor
    trend?: 'positive' | 'negative'
}
export type CardStatsType = {
    statsVertical: CardStatsVerticalType[]
    statsCharacter: CardStatsCharacterType[]
    statsHorizontal: CardStatsHorizontalType[]
}